requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via py-cord
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via pynacl
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.16.1
    # via google-api-python-client
google-api-python-client==2.116.0
    # via -r requirements.in
google-auth==2.27.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.62.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.6
    # via
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
protobuf==4.25.2
    # via
    #   google-api-core
    #   googleapis-common-protos
py-cord[voice]==2.4.1
    # via
    #   -r requirements.in
    #   py-cord
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via py-cord
pyparsing==3.1.1
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
rsa==4.9
    # via google-auth
tweepy==4.14.0
    # via -r requirements.in
typing-extensions==4.9.0
    # via py-cord
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.0
    # via requests
yarl==1.9.4
    # via aiohttp
